+++ src/lti_1p3_provider/tests/test_views.py
@@ -12,6 +12,7 @@ from urllib import parse
 import jwt
 import pytest
 from bs4 import BeautifulSoup
+from common.djangoapps.student.tests.factories import UserProfileFactory
 from crum import CurrentRequestUserMiddleware
 from django.conf import settings
 from django.contrib.auth.models import AnonymousUser
@@ -29,7 +30,11 @@ from lti_1p3_provider.api.ssl_services import (
 )
 from lti_1p3_provider.models import LtiGradedResource, LtiProfile
 from lti_1p3_provider.session_access import LTI_SESSION_KEY
-from lti_1p3_provider.views import DisplayTargetResource, LtiToolLaunchView
+from lti_1p3_provider.views import (
+    LTI_1P3_EMAIL_META_KEY,
+    DisplayTargetResource,
+    LtiToolLaunchView,
+)
 
 from . import factories, fakes
 from .base import URL_LIB_LTI_JWKS
@@ -194,6 +199,136 @@ class TestLtiToolLaunchView:
 
         assert resp.status_code == 404
 
+    def test_successful_launch_with_email_sets_email_in_profile(self, client):
+        """If email claim provided, sets it in the LtiProfile and UserProfile
+
+        User does not yet exist, so User, LtiProfile, and UserProfile are created
+        """
+        email = "test@example.com"
+        target_link_uri = _get_target_link_uri(
+            str(factories.COURSE_KEY), str(factories.USAGE_KEY)
+        )
+        id_token = factories.IdTokenFactory(
+            aud=self.tool.client_id,
+            nonce="nonce",
+            target_link_uri=target_link_uri,
+            email=email,
+        )
+        encoded = _encode_platform_jwt(id_token, self.kid)
+        payload = {"state": "state", "id_token": encoded}
+
+        resp = client.post(self.launch_endpoint, payload)
+
+        assert resp.status_code == 302
+        redirect_uri = reverse(
+            "lti_1p3_provider:lti-display",
+            kwargs={
+                "course_id": str(factories.COURSE_KEY),
+                "usage_id": str(factories.USAGE_KEY),
+            },
+        )
+        assert resp.url == f"http://localhost{redirect_uri}"
+        lti_profile = LtiProfile.objects.get_from_claims(
+            iss=id_token["iss"], aud=id_token["aud"], sub=id_token["sub"]
+        )
+        assert lti_profile.email == email
+        assert lti_profile.user.profile.get_meta()[LTI_1P3_EMAIL_META_KEY] == email
+
+    def test_existing_user_profile_with_no_email_gets_updated(self, client):
+        """If email claim provided, updates the existing LtiProfile and UserProfile
+
+        User, LtiProfile, and UserProfile exist, but UserProfile has no email
+        """
+        email = "test@example.com"
+        payload = self._get_payload(factories.COURSE_KEY, factories.USAGE_KEY)
+        target_link_uri = _get_target_link_uri(
+            str(factories.COURSE_KEY), str(factories.USAGE_KEY)
+        )
+        # This id token has an email
+        id_token = factories.IdTokenFactory(
+            aud=self.tool.client_id,
+            nonce="nonce",
+            target_link_uri=target_link_uri,
+            email="test@example.com",
+        )
+        encoded = _encode_platform_jwt(id_token, self.kid)
+        payload = {"state": "state", "id_token": encoded}
+        # Create a profile with no email
+        lti_profile = LtiProfile.objects.get_or_create_from_claims(
+            iss=id_token["iss"], aud=id_token["aud"], sub=id_token["sub"], email=""
+        )
+        # UserProfile exists, but has no email
+        UserProfileFactory(user=lti_profile.user)
+
+        resp = client.post(self.launch_endpoint, payload)
+
+        assert resp.status_code == 302
+        redirect_uri = reverse(
+            "lti_1p3_provider:lti-display",
+            kwargs={
+                "course_id": str(factories.COURSE_KEY),
+                "usage_id": str(factories.USAGE_KEY),
+            },
+        )
+        assert resp.url == f"http://localhost{redirect_uri}"
+        fetched_profile = LtiProfile.objects.get_from_claims(
+            iss=id_token["iss"], aud=id_token["aud"], sub=id_token["sub"]
+        )
+        assert fetched_profile == lti_profile
+        assert fetched_profile.email == email
+        assert fetched_profile.user.profile.get_meta()[LTI_1P3_EMAIL_META_KEY] == email
+
+    def test_existing_user_profile_with_email_gets_updated(self, client):
+        """If email claim provided, updates the existing LtiProfile and UserProfile
+
+        User, LtiProfile, and UserProfile exist, and UserProfile has email set
+        """
+        email = "test@example.com"
+        payload = self._get_payload(factories.COURSE_KEY, factories.USAGE_KEY)
+        target_link_uri = _get_target_link_uri(
+            str(factories.COURSE_KEY), str(factories.USAGE_KEY)
+        )
+        # This id token has an email
+        id_token = factories.IdTokenFactory(
+            aud=self.tool.client_id,
+            nonce="nonce",
+            target_link_uri=target_link_uri,
+            email="test@example.com",
+        )
+        encoded = _encode_platform_jwt(id_token, self.kid)
+        payload = {"state": "state", "id_token": encoded}
+        # Create a profile with email already set
+        lti_profile = LtiProfile.objects.get_or_create_from_claims(
+            iss=id_token["iss"],
+            aud=id_token["aud"],
+            sub=id_token["sub"],
+            email="already@set.com",
+        )
+        # UserProfile exists, and has email set
+        user_profile = UserProfileFactory(
+            user=lti_profile.user,
+            meta=f'{{"{LTI_1P3_EMAIL_META_KEY}": "already@set.com"}}',
+        )
+        assert user_profile.get_meta()[LTI_1P3_EMAIL_META_KEY] == "already@set.com"
+
+        resp = client.post(self.launch_endpoint, payload)
+
+        assert resp.status_code == 302
+        redirect_uri = reverse(
+            "lti_1p3_provider:lti-display",
+            kwargs={
+                "course_id": str(factories.COURSE_KEY),
+                "usage_id": str(factories.USAGE_KEY),
+            },
+        )
+        assert resp.url == f"http://localhost{redirect_uri}"
+        fetched_profile = LtiProfile.objects.get_from_claims(
+            iss=id_token["iss"], aud=id_token["aud"], sub=id_token["sub"]
+        )
+        assert fetched_profile == lti_profile
+        assert fetched_profile.email == email
+        assert fetched_profile.user.profile.get_meta()[LTI_1P3_EMAIL_META_KEY] == email
+
     def test_successful_launch_no_gate(self, client):
         """Test successsful launch with no gate in place"""
         payload = self._get_payload(factories.COURSE_KEY, factories.USAGE_KEY)
@@ -209,6 +344,10 @@ class TestLtiToolLaunchView:
             },
         )
         assert resp.url == f"http://localhost{redirect_uri}"
+        # Since no email in id_token, the profile should have no email
+        fetched_profile = LtiProfile.objects.first()
+        assert fetched_profile.email == ""
+        assert LTI_1P3_EMAIL_META_KEY not in fetched_profile.user.profile.get_meta()
 
     def test_successful_launch_with_gate(self, client):
         """Test successful launch where target_link_uri is allowed by gate"""